pyproject.toml
[project]
name = "fastedit"
version = "0.1.0"
description = "AST-aware code editing — 45% fewer output tokens than unified diffs"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-c>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    "tree-sitter-swift",
    "tree-sitter-kotlin>=1.0.0",
    "tree-sitter-c-sharp>=0.23.0",
    "tree-sitter-php>=0.23.0",
    "openai>=1.50.0",
    "rich>=13.0.0",
    "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.18.0",
    "mlx-lm>=0.18.0",
]
vllm = [
    "vllm>=0.6.0",
]
mcp = [
    "fastmcp>=0.4.0",
]

[project.scripts]
fastedit = "fastedit.cli:main"
fastedit-hook = "fastedit.hook:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastedit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
